Refactor budget repository

+++ app/src/main/java/com/udinesfata/expenz/data/datasource/local/BudgetLocalDataSource.kt
@@ -2,6 +2,8 @@ package com.udinesfata.expenz.data.datasource.local
 
 import com.udinesfata.expenz.data.datasource.local.database.BudgetDao
 import com.udinesfata.expenz.data.model.local.BudgetDb
+import com.udinesfata.expenz.data.model.local.SYNC_OPERATION_DELETE
+import com.udinesfata.expenz.data.model.query.BudgetQuery
 
 class BudgetLocalDataSource(
     private val budgetDao: BudgetDao
@@ -10,15 +12,37 @@ class BudgetLocalDataSource(
         return budgetDao.getBudget(id)
     }
 
+    suspend fun getBudgets(query: BudgetQuery): List<BudgetDb> {
+        return budgetDao.getBudgets()
+    }
+
     suspend fun createBudget(budget: BudgetDb) {
         budgetDao.createBudget(budget)
+    }
+
+    suspend fun createBudgets(budgets: List<BudgetDb>) {
+        for (budget in budgets) {
+            budgetDao.createBudget(budget)
+        }
     }
 
     suspend fun updateBudget(budget: BudgetDb) {
         budgetDao.updateBudget(budget)
     }
 
-    suspend fun deleteBudget(id: Int) {
-        budgetDao.deleteBudget(id)
+    suspend fun deleteBudget(id: Int, flagOnly: Boolean = false) {
+        if (!flagOnly) {
+            budgetDao.deleteBudget(id)
+        } else {
+            val budget = budgetDao.getBudget(id)
+            budget?.let {
+                budgetDao.updateBudget(
+                    it.copy(
+                        isSynced = false,
+                        syncOperation = SYNC_OPERATION_DELETE
+                    )
+                )
+            }
+        }
     }
 }
